+++ CHANGELOG.md
@@ -8,6 +8,8 @@ _These will be removed in the next major release_
 
 ## 0.27.0
 - Added `Add-BitbucketRepositoryBranch` to create a new branch in a repo
+- Added `Get-BitbucketRepositoryBranchModel` to get Branching Model in a repo
+- Added `Set-BitbucketRepositoryBranchModel` to modify Branching Model in a repo
 - Added `New-BitbucketRepositoryBranchRestrictionPermissionCheck`
 - Updated `New-BitbucketRepositoryBranchRestrictionMergeCheck` with missing merge checks
 - Updated BranchRestriction Class to support branching_model matching type